process-api/requirements.txt
aniso8601==8.0.0
attrs==20.2.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-RESTful==0.3.8
grpcio==1.32.0
grpcio-tools==1.32.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
numpy==1.18.5
opencv-python==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
zipp==3.2.0
